+++ crates/rtlola-interpreter/src/evaluator.rs
@@ -2,18 +2,21 @@ use std::cell::RefCell;
 use std::collections::HashMap;
 use std::ops::Not;
 use std::rc::Rc;
+use std::time::Duration;
 
 use bit_set::BitSet;
 use itertools::Itertools;
+use num::traits::Inv;
 use rtlola_frontend::mir::{
-    ActivationCondition as Activation, InputReference, OutputReference, PacingType, RtLolaMir, Stream, StreamReference,
-    Task, TimeDrivenStream, Trigger, WindowReference,
+    ActivationCondition as Activation, InputReference, OutputKind, OutputReference, PacingLocality, PacingType,
+    RtLolaMir, Stream, StreamReference, Task, TimeDrivenStream, Trigger, WindowReference,
 };
-use string_template::Template;
+use uom::si::rational64::Time as UOM_Time;
+use uom::si::time::nanosecond;
 
 use crate::api::monitor::{Change, Instance};
 use crate::closuregen::{CompiledExpr, Expr};
-use crate::monitor::Tracer;
+use crate::monitor::{Parameters, Tracer};
 use crate::schedule::{DynamicSchedule, EvaluationTask};
 use crate::storage::{GlobalStore, InstanceAggregationTrait, InstanceStore, Value, WindowParameterization};
 use crate::Time;
@@ -44,7 +47,6 @@ pub(crate) struct EvaluatorData {
     fresh_triggers: BitSet,
     triggers: Vec<Option<Trigger>>,
     time_driven_streams: Vec<Option<TimeDrivenStream>>,
-    trigger_templates: Vec<Option<Template>>,
     closing_streams: Vec<OutputReference>,
     ir: RtLolaMir,
     dyn_schedule: Rc<RefCell<DynamicSchedule>>,
@@ -82,7 +84,6 @@ pub(crate) struct Evaluator {
     // Indexed by output reference
     time_driven_streams: &'static [Option<TimeDrivenStream>],
 
-    trigger_templates: &'static [Option<Template>],
     closing_streams: &'static [OutputReference],
     ir: &'static RtLolaMir,
     dyn_schedule: &'static RefCell<DynamicSchedule>,
@@ -112,7 +113,7 @@ impl EvaluatorData {
             .iter()
             .map(|o| {
                 match &o.eval.eval_pacing {
-                    PacingType::Periodic(_) => ActivationConditionOp::TimeDriven,
+                    PacingType::GlobalPeriodic(_) | PacingType::LocalPeriodic(_) => ActivationConditionOp::TimeDriven,
                     PacingType::Event(ac) => ActivationConditionOp::new(ac, ir.inputs.len()),
                     PacingType::Constant => ActivationConditionOp::True,
                 }
@@ -123,7 +124,7 @@ impl EvaluatorData {
             .iter()
             .map(|o| {
                 match &o.spawn.pacing {
-                    PacingType::Periodic(_) => ActivationConditionOp::TimeDriven,
+                    PacingType::GlobalPeriodic(_) | PacingType::LocalPeriodic(_) => ActivationConditionOp::TimeDriven,
                     PacingType::Event(ac) => ActivationConditionOp::new(ac, ir.inputs.len()),
                     PacingType::Constant => ActivationConditionOp::True,
                 }
@@ -134,7 +135,7 @@ impl EvaluatorData {
             .iter()
             .map(|o| {
                 match &o.close.pacing {
-                    PacingType::Periodic(_) => ActivationConditionOp::TimeDriven,
+                    PacingType::GlobalPeriodic(_) | PacingType::LocalPeriodic(_) => ActivationConditionOp::TimeDriven,
                     PacingType::Event(ac) => ActivationConditionOp::new(ac, ir.inputs.len()),
                     PacingType::Constant => ActivationConditionOp::True,
                 }
@@ -163,16 +164,12 @@ impl EvaluatorData {
             .into_group_map();
 
         for t in &ir.triggers {
-            triggers[t.reference.out_ix()] = Some(t.clone());
+            triggers[t.output_reference.out_ix()] = Some(*t);
         }
         let mut time_driven_streams = vec![None; ir.outputs.len()];
         for t in &ir.time_driven {
             time_driven_streams[t.reference.out_ix()] = Some(*t);
         }
-        let trigger_templates = triggers
-            .iter()
-            .map(|t| t.as_ref().map(|t| Template::new(&t.message)))
-            .collect();
         EvaluatorData {
             layers,
             stream_activation_conditions: stream_acs,
@@ -188,7 +185,6 @@ impl EvaluatorData {
             fresh_triggers,
             triggers,
             time_driven_streams,
-            trigger_templates,
             closing_streams,
             ir,
             dyn_schedule,
@@ -290,7 +286,6 @@ impl EvaluatorData {
             fresh_triggers: &mut leaked_data.fresh_triggers,
             triggers: &leaked_data.triggers,
             time_driven_streams: &leaked_data.time_driven_streams,
-            trigger_templates: &leaked_data.trigger_templates,
             closing_streams: &leaked_data.closing_streams,
             ir: &leaked_data.ir,
             dyn_schedule: &leaked_data.dyn_schedule,
@@ -357,6 +352,23 @@ impl Evaluator {
                     vec![]
                 };
                 changes.is_empty().not().then(|| (o.reference.out_ix(), changes))
+            })
+            .collect()
+    }
+
+    /// NOT for external use because the values are volatile
+    pub(crate) fn peek_violated_triggers_messages(&self) -> Vec<(OutputReference, Parameters, String)> {
+        self.peek_fresh_outputs()
+            .into_iter()
+            .filter(|(o_ref, _)| matches!(self.ir.outputs[*o_ref].kind, OutputKind::Trigger(_)))
+            .flat_map(|(o_ref, changes)| {
+                changes.into_iter().filter_map(move |change| {
+                    match change {
+                        Change::Value(parameters, Value::Str(msg)) => Some((o_ref, parameters, msg.into())),
+                        Change::Value(_, _) => unreachable!("trigger values are strings; checked by the frontend"),
+                        _ => None,
+                    }
+                })
             })
             .collect()
     }
@@ -570,13 +582,22 @@ impl Evaluator {
         // Schedule instance evaluation if stream is periodic
         if let Some(tds) = self.time_driven_streams[output] {
             let mut schedule = (*self.dyn_schedule).borrow_mut();
-            schedule.schedule_evaluation(output, parameter_values.as_slice(), ts, tds.period_in_duration());
 
-            // Schedule close if it depends on current instance
-            if stream.close.has_self_reference {
-                // we have a synchronous access to self -> period of close should be the same as og self
-                debug_assert!(matches!(&stream.close.pacing, PacingType::Periodic(f) if *f == tds.frequency));
-                schedule.schedule_close(output, parameter_values.as_slice(), ts, tds.period_in_duration());
+            // Schedule eval if it has local pacing
+            if tds.locality == PacingLocality::Local {
+                schedule.schedule_evaluation(output, parameter_values.as_slice(), ts, tds.period_in_duration());
+            }
+
+            // Schedule close if it has local pacing
+            if let PacingType::LocalPeriodic(f) = stream.close.pacing {
+                let period = Duration::from_nanos(
+                    UOM_Time::new::<uom::si::time::second>(f.get::<uom::si::frequency::hertz>().inv())
+                        .get::<nanosecond>()
+                        .to_integer()
+                        .try_into()
+                        .expect("Period [ns] too large for u64!"),
+                );
+                schedule.schedule_close(output, parameter_values.as_slice(), ts, period);
             }
         }
     }
@@ -666,8 +687,15 @@ impl Evaluator {
             schedule.remove_evaluation(output, parameter, tds.period_in_duration());
 
             // Remove close from schedule if it depends on current instance
-            if stream.close.has_self_reference {
-                schedule.remove_close(output, parameter, tds.period_in_duration());
+            if let PacingType::LocalPeriodic(f) = stream.close.pacing {
+                let period = Duration::from_nanos(
+                    UOM_Time::new::<uom::si::time::second>(f.get::<uom::si::frequency::hertz>().inv())
+                        .get::<nanosecond>()
+                        .to_integer()
+                        .try_into()
+                        .expect("Period [ns] too large for u64!"),
+                );
+                schedule.remove_close(output, parameter, period);
             }
         }
     }
@@ -699,25 +727,29 @@ impl Evaluator {
         }
     }
 
+    fn eval_stream_instances(&mut self, output: OutputReference, ts: Time, tracer: &mut impl Tracer) {
+        if self.ir.output(StreamReference::Out(output)).is_parameterized() {
+            let parameter: Vec<Vec<Value>> = self
+                .global_store
+                .get_out_instance_collection(output)
+                .all_parameter()
+                .cloned()
+                .collect();
+            for instance in parameter {
+                tracer.instance_eval_start(output, instance.as_slice());
+                self.eval_stream_instance(output, instance.as_slice(), ts);
+                tracer.instance_eval_end(output, instance.as_slice());
+            }
+        } else if self.global_store.get_out_instance(output).is_active() {
+            tracer.instance_eval_start(output, &[]);
+            self.eval_stream_instance(output, &[], ts);
+            tracer.instance_eval_end(output, &[]);
+        }
+    }
+
     fn eval_event_driven_output(&mut self, output: OutputReference, ts: Time, tracer: &mut impl Tracer) {
         if self.stream_activation_conditions[output].eval(self.fresh_inputs) {
-            if self.ir.output(StreamReference::Out(output)).is_parameterized() {
-                let parameter: Vec<Vec<Value>> = self
-                    .global_store
-                    .get_out_instance_collection(output)
-                    .all_parameter()
-                    .cloned()
-                    .collect();
-                for instance in parameter {
-                    tracer.instance_eval_start(output, instance.as_slice());
-                    self.eval_stream_instance(output, instance.as_slice(), ts);
-                    tracer.instance_eval_end(output, instance.as_slice());
-                }
-            } else if self.global_store.get_out_instance(output).is_active() {
-                tracer.instance_eval_start(output, &[]);
-                self.eval_stream_instance(output, &[], ts);
-                tracer.instance_eval_end(output, &[]);
-            }
+            self.eval_stream_instances(output, ts, tracer)
         }
     }
 
@@ -734,6 +766,9 @@ impl Evaluator {
                     tracer.instance_eval_start(idx, parameter.as_slice());
                     self.eval_stream_instance(idx, parameter.as_slice(), ts);
                     tracer.instance_eval_end(idx, parameter.as_slice());
+                },
+                EvaluationTask::EvaluateInstances(idx) => {
+                    self.eval_stream_instances(idx, ts, tracer);
                 },
                 EvaluationTask::Spawn(idx) => {
                     tracer.spawn_start(idx);
@@ -775,40 +810,6 @@ impl Evaluator {
         }
     }
 
-    /// Creates the current trigger message by substituting the format placeholders with he current values of the info streams.
-    /// NOT for external use because the values are volatile
-    pub(crate) fn format_trigger_message(&self, trigger_ref: OutputReference) -> String {
-        let trigger = self
-            .is_trigger(trigger_ref)
-            .expect("Output reference must refer to a trigger");
-        let values: Vec<String> = trigger
-            .info_streams
-            .iter()
-            .map(|sr| {
-                self.peek_value(*sr, &[], 0)
-                    .map_or("None".to_string(), |v| v.to_string())
-            })
-            .collect();
-        let args: Vec<&str> = values.iter().map(|s| s.as_str()).collect();
-        self.trigger_templates[trigger_ref]
-            .as_ref()
-            .expect("Output reference must refer to a trigger")
-            .render_positional(args.as_slice())
-    }
-
-    /// Return the current values of the info streams
-    /// NOT for external use because the values are volatile
-    pub(crate) fn peek_info_stream_values(&self, trigger_ref: OutputReference) -> Vec<Option<Value>> {
-        let trigger = self
-            .is_trigger(trigger_ref)
-            .expect("Output reference must refer to a trigger");
-        trigger
-            .info_streams
-            .iter()
-            .map(|sr| self.peek_value(*sr, &[], 0))
-            .collect()
-    }
-
     fn eval_stream_instance(&mut self, output: OutputReference, parameter: &[Value], ts: Time) {
         let ix = output;
 
@@ -836,10 +837,7 @@ impl Evaluator {
         self.fresh_outputs.insert(ix);
 
         if self.is_trigger(output).is_some() {
-            // Check if we have to emit a trigger.
-            if let Value::Bool(true) = res.clone() {
-                self.fresh_triggers.insert(ix);
-            }
+            self.fresh_triggers.insert(ix);
         }
 
         // Update Instance aggregations